+++ SDialogs/src/main/java/smith/lib/alerts/dialog/callbacks/OnProgressCallback.java
@@ -19,12 +19,18 @@
 package smith.lib.alerts.dialog.callbacks;
 
 /**
- * On dismiss callback useful for dismissing SDialogs.
+ * on progress changed callback for ProgressSDialog.
  */
-@FunctionalInterface
-public interface OnDismissCallBack {
+public interface OnProgressCallback {
     /**
-     * Triggered whenever the SDialogs are dismissed.
+     * triggered when progress is continuing and returns:
+     *
+     * @param progress the current progress as int.
+     * @param percent  the progress percent from 100% as int.
      */
-    void onDismiss();
+    void onProgress(int progress, int percent);
+    /**
+     * triggered as far as the progress is completed.
+     */
+    void onFinish();
 }